--- STTstart.py
@@ -1,103 +1,0 @@
-import requests
-import json
-import os
-
-audio_path = './audio/'
-text_path = './text/org/'
-audio_flie_list = os.listdir(audio_path)
-
-
-class ClovaSpeechClient:
-    # Clova Speech invoke URL(have to change key)
-    invoke_url = 'https://clovaspeech-gw.ncloud.com/external/v1/2718/92ca14e89f0ed52a426f0fb7889839f4678503352f0691e534d0b769d6db6664'
-    # Clova Speech secret key(have to change key)
-    secret = 'bd99361c10f74126a4c63222053f9dd8'
-
-    def req_url(self, url, completion, callback=None, userdata=None, forbiddens=None, boostings=None, wordAlignment=True, fullText=True, diarization=None):
-        request_body = {
-            'url': url,
-            'language': 'ko-KR',
-            'completion': completion,
-            'callback': callback,
-            'userdata': userdata,
-            'wordAlignment': wordAlignment,
-            'fullText': fullText,
-            'forbiddens': forbiddens,
-            'boostings': boostings,
-            'diarization': diarization,
-        }
-        headers = {
-            'Accept': 'application/json;UTF-8',
-            'Content-Type': 'application/json;UTF-8',
-            'X-CLOVASPEECH-API-KEY': self.secret
-        }
-        return requests.post(headers=headers,
-                             url=self.invoke_url + '/recognizer/url',
-                             data=json.dumps(request_body).encode('UTF-8'))
-
-    def req_object_storage(self, data_key, completion, callback=None, userdata=None, forbiddens=None, boostings=None,
-                           wordAlignment=True, fullText=True, diarization=None):
-        request_body = {
-            'dataKey': data_key,
-            'language': 'ko-KR',
-            'completion': completion,
-            'callback': callback,
-            'userdata': userdata,
-            'wordAlignment': wordAlignment,
-            'fullText': fullText,
-            'forbiddens': forbiddens,
-            'boostings': boostings,
-            'diarization': diarization,
-        }
-        headers = {
-            'Accept': 'application/json;UTF-8',
-            'Content-Type': 'application/json;UTF-8',
-            'X-CLOVASPEECH-API-KEY': self.secret
-        }
-        return requests.post(headers=headers,
-                             url=self.invoke_url + '/recognizer/object-storage',
-                             data=json.dumps(request_body).encode('UTF-8'))
-
-    def req_upload(self, file, completion, callback=None, userdata=None, forbiddens=None, boostings=None,
-                   wordAlignment=True, fullText=True, diarization=None):
-        request_body = {
-            'language': 'ko-KR',
-            'completion': completion,
-            'callback': callback,
-            'userdata': userdata,
-            'wordAlignment': wordAlignment,
-            'fullText': fullText,
-            'forbiddens': forbiddens,
-            'boostings': boostings,
-            'diarization': diarization,
-        }
-        headers = {
-            'Accept': 'application/json;UTF-8',
-            'X-CLOVASPEECH-API-KEY': self.secret
-        }
-        print(json.dumps(request_body, ensure_ascii=False).encode('UTF-8'))
-        files = {
-            'media': open(file, 'rb'),
-            'params': (None, json.dumps(request_body, ensure_ascii=False).encode('UTF-8'), 'application/json')
-        }
-        response = requests.post(headers=headers, url=self.invoke_url + '/recognizer/upload', files=files)
-        return response
-
-if __name__ == '__main__':
-
-    for i in audio_flie_list:
-
-        # res = ClovaSpeechClient().req_url(url='http://example.com/media.mp3', completion='sync')
-        # res = ClovaSpeechClient().req_object_storage(data_key='data/media.mp3', completion='sync')
-        
-        res = ClovaSpeechClient().req_upload(file = audio_path + i, completion='sync')
-        json_temp = res.json()
-        res_txt = json_temp.get("text")
-
-        f = open(text_path + i[:-4] + '.txt', 'w')
-        f.write(str(res_txt))
-        f.close()
-
-        os.system('rm ' + audio_path + i)
-
-        #print(res.text)
