+++ src/stopwords.txt
@@ -1,0 +1,1 @@
+그래서
