+++ text/sum/sample_1_sum.txt
@@ -1,0 +1,7 @@
+------- keywords -------
+기침
+
+
+------- summary -------
+코로나19 예방 수칙입니다.
+손을 자주 씻기 마스크 착용하기 기침할 땐 입과 코 가리기 발열 기침 인후통 등 증상 의심 시에는 1339 또는 보건소와 상담하시기 바랍니다.
